Refactored syntax-checking
Moved syntax-checking inside the Runner class in order to avoid cyclic
loading issues.
