+++ src/ansiblelint/rules/syntax_check.py
@@ -1,22 +1,10 @@
 """Rule definition for ansible syntax check."""
 from __future__ import annotations
 
-import json
 import re
-import subprocess
-import sys
-import tempfile
 from dataclasses import dataclass
-from typing import Any
 
-from ansiblelint._internal.rules import BaseRule, RuntimeErrorRule
-from ansiblelint.app import App, get_app
-from ansiblelint.config import options
-from ansiblelint.errors import MatchError
-from ansiblelint.file_utils import Lintable
-from ansiblelint.logger import timed_info
 from ansiblelint.rules import AnsibleLintRule
-from ansiblelint.text import strip_ansi_escape
 
 
 @dataclass
@@ -68,203 +56,3 @@ class AnsibleSyntaxCheckRule(AnsibleLintRule):
     tags = ["core", "unskippable"]
     version_added = "v5.0.0"
     _order = 0
-
-    @staticmethod
-    # pylint: disable=too-many-locals
-    def _get_ansible_syntax_check_matches(
-        lintable: Lintable,
-        app: App,
-    ) -> list[MatchError]:
-        """Run ansible syntax check and return a list of MatchError(s)."""
-        default_rule: BaseRule = AnsibleSyntaxCheckRule()
-        fh = None
-        results = []
-        if lintable.kind not in ("playbook", "role"):
-            return []
-
-        with timed_info(
-            "Executing syntax check on %s %s",
-            lintable.kind,
-            lintable.path,
-        ):
-            if lintable.kind == "role":
-                playbook_text = f"""
----
-- name: Temporary playbook for role syntax check
-  hosts: localhost
-  tasks:
-    - ansible.builtin.import_role:
-        name: {lintable.path.expanduser()!s}
-"""
-                # pylint: disable=consider-using-with
-                fh = tempfile.NamedTemporaryFile(mode="w", suffix=".yml", prefix="play")
-                fh.write(playbook_text)
-                fh.flush()
-                playbook_path = fh.name
-            else:
-                playbook_path = str(lintable.path.expanduser())
-            # To avoid noisy warnings we pass localhost as current inventory:
-            # [WARNING]: No inventory was parsed, only implicit localhost is available
-            # [WARNING]: provided hosts list is empty, only localhost is available. Note that the implicit localhost does not match 'all'
-            cmd = [
-                "ansible-playbook",
-                "-i",
-                "localhost,",
-                "--syntax-check",
-                playbook_path,
-            ]
-            if options.extra_vars:
-                cmd.extend(["--extra-vars", json.dumps(options.extra_vars)])
-
-            # To reduce noisy warnings like
-            # CryptographyDeprecationWarning: Blowfish has been deprecated
-            # https://github.com/paramiko/paramiko/issues/2038
-            env = app.runtime.environ.copy()
-            env["PYTHONWARNINGS"] = "ignore"
-
-            run = subprocess.run(
-                cmd,
-                stdin=subprocess.PIPE,
-                capture_output=True,
-                shell=False,  # needed when command is a list
-                text=True,
-                check=False,
-                env=env,
-            )
-
-        if run.returncode != 0:
-            message = None
-            filename = lintable
-            lineno = 1
-            column = None
-
-            stderr = strip_ansi_escape(run.stderr)
-            stdout = strip_ansi_escape(run.stdout)
-            if stderr:
-                details = stderr
-                if stdout:
-                    details += "\n" + stdout
-            else:
-                details = stdout
-
-            for pattern in OUTPUT_PATTERNS:
-                rule = default_rule
-                match = re.search(pattern.regex, stderr)
-                if match:
-                    groups = match.groupdict()
-                    title = groups.get("title", match.group(0))
-                    details = groups.get("details", "")
-                    lineno = int(groups.get("line", 1))
-
-                    if "filename" in groups:
-                        filename = Lintable(groups["filename"])
-                    else:
-                        filename = lintable
-                    column = int(groups.get("column", 1))
-                    results.append(
-                        MatchError(
-                            message=title,
-                            lintable=filename,
-                            lineno=lineno,
-                            column=column,
-                            rule=rule,
-                            details=details,
-                            tag=f"{rule.id}[{pattern.tag}]",
-                        ),
-                    )
-
-            if not results:
-                rule = RuntimeErrorRule()
-                message = (
-                    f"Unexpected error code {run.returncode} from "
-                    f"execution of: {' '.join(cmd)}"
-                )
-                results.append(
-                    MatchError(
-                        message=message,
-                        lintable=filename,
-                        lineno=lineno,
-                        column=column,
-                        rule=rule,
-                        details=details,
-                        tag="",
-                    ),
-                )
-
-        if fh:
-            fh.close()
-        return results
-
-
-# testing code to be loaded only with pytest or when executed the rule file
-if "pytest" in sys.modules:
-
-    def test_get_ansible_syntax_check_matches() -> None:
-        """Validate parsing of ansible output."""
-        lintable = Lintable(
-            "examples/playbooks/conflicting_action.yml",
-            kind="playbook",
-        )
-        # pylint: disable=protected-access
-        result = AnsibleSyntaxCheckRule._get_ansible_syntax_check_matches(
-            lintable,
-            app=get_app(),
-        )
-        assert result[0].lineno == 4
-        assert result[0].column == 7
-        assert (
-            result[0].message
-            == "conflicting action statements: ansible.builtin.debug, ansible.builtin.command"
-        )
-        # We internally convert absolute paths returned by ansible into paths
-        # relative to current directory.
-        assert result[0].filename.endswith("/conflicting_action.yml")
-        assert len(result) == 1
-
-    def test_empty_playbook() -> None:
-        """Validate detection of empty-playbook."""
-        lintable = Lintable("examples/playbooks/empty_playbook.yml", kind="playbook")
-        # pylint: disable=protected-access
-        result = AnsibleSyntaxCheckRule._get_ansible_syntax_check_matches(
-            lintable,
-            app=get_app(),
-        )
-        assert result[0].lineno == 1
-        # We internally convert absolute paths returned by ansible into paths
-        # relative to current directory.
-        assert result[0].filename.endswith("/empty_playbook.yml")
-        assert result[0].tag == "syntax-check[empty-playbook]"
-        assert result[0].message == "Empty playbook, nothing to do"
-        assert len(result) == 1
-
-    def test_extra_vars_passed_to_command(config_options: Any) -> None:
-        """Validate `extra-vars` are passed to syntax check command."""
-        config_options.extra_vars = {
-            "foo": "bar",
-            "complex_variable": ":{;\t$()",
-        }
-        lintable = Lintable("examples/playbooks/extra_vars.yml", kind="playbook")
-
-        # pylint: disable=protected-access
-        result = AnsibleSyntaxCheckRule._get_ansible_syntax_check_matches(
-            lintable,
-            app=get_app(),
-        )
-
-        assert not result
-
-    def test_syntax_check_role() -> None:
-        """Validate syntax check of a broken role."""
-        # pylint: disable=import-outside-toplevel
-        from ansiblelint.rules import RulesCollection
-        from ansiblelint.runner import Runner
-
-        lintable = Lintable("examples/playbooks/roles/invalid_due_syntax", kind="role")
-        # pylint: disable=protected-access
-        rules = RulesCollection()
-        result = Runner(lintable, rules=rules).run()
-        assert len(result) == 1, result
-        assert result[0].lineno == 2
-        assert result[0].filename == "examples/roles/invalid_due_syntax/tasks/main.yml"
-        assert result[0].tag == "syntax-check[specific]"
-        assert result[0].message == "no module/action detected in task."
